refactor: 패키지 분리 및 302 redirect
