+++ tomcat/src/main/java/org/apache/coyote/http11/component/common/Body.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.component;
+package org.apache.coyote.http11.component.common;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
