+++ tomcat/src/main/java/org/apache/coyote/http11/component/common/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.component;
+package org.apache.coyote.http11.component.common;
 
 public class StatusCode {
