+++ tomcat/src/main/java/org/apache/coyote/http11/component/common/Version.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.component;
+package org.apache.coyote.http11.component.common;
 
 import java.util.List;
