+++ tomcat/src/test/java/org/apache/coyote/http11/component/request/HttpRequestTest.java
@@ -2,7 +2,6 @@ package org.apache.coyote.http11.component.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.coyote.http11.component.HttpRequest;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
