+++ tomcat/src/test/java/org/apache/coyote/http11/component/request/MethodTest.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.component.request;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import org.apache.coyote.http11.component.Method;
+import org.apache.coyote.http11.component.common.Method;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
