+++ tomcat/src/test/java/org/apache/coyote/http11/component/request/RequestLineTest.java
@@ -3,7 +3,6 @@ package org.apache.coyote.http11.component.request;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import org.apache.coyote.http11.component.RequestLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
