+++ tomcat/src/test/java/org/apache/coyote/http11/component/request/VersionTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import org.apache.coyote.http11.component.Version;
+import org.apache.coyote.http11.component.common.Version;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
